refactor(town): extract talk handler to talk.js and delegate from runtime.js
