+++ ROguelike 3_10_2025 filut jarjestyksessa luultavasti vikaa/core/town/talk.js
@@ -1,0 +1,431 @@
+import { getMod } from "../../utils/access.js";
+
+/**
+ * Town talk / NPC interaction logic extracted from core/town/runtime.js.
+ *
+ * Exported as talk(ctx, bumpAtX?, bumpAtY?) and used by TownRuntime.
+ */
+export function talk(ctx, bumpAtX = null, bumpAtY = null) {
+  if (ctx.mode !== "town") return false;
+  const npcs = ctx.npcs || [];
+  const near = [];
+  for (const n of npcs) {
+    const d = Math.abs(n.x - ctx.player.x) + Math.abs(n.y - ctx.player.y);
+    if (d <= 1) near.push(n);
+  }
+  if (!near.length) {
+    ctx.log && ctx.log("There is no one to talk to here.");
+    return false;
+  }
+
+  // Prefer the NPC occupying the attempted bump tile if provided,
+  // otherwise prefer a shopkeeper among adjacent NPCs, otherwise pick randomly.
+  let npc = null;
+  if (typeof bumpAtX === "number" && typeof bumpAtY === "number") {
+    npc = near.find(n => n.x === bumpAtX && n.y === bumpAtY) || null;
+  }
+  if (!npc) {
+    npc = near.find(n => (n.isShopkeeper || n._shopRef)) || null;
+  }
+  const pick = (arr, rng) => {
+    try {
+      const RU = (typeof window !== "undefined") ? window.RNGUtils : null;
+      if (RU && typeof RU.int === "function") {
+        const rfn =
+          typeof rng === "function"
+            ? rng
+            : typeof ctx.rng === "function"
+            ? ctx.rng
+            : undefined;
+        if (typeof rfn === "function") {
+          const idx = RU.int(0, arr.length - 1, rfn);
+          return arr[idx] || arr[0];
+        }
+      }
+    } catch (_) {}
+    if (typeof rng === "function") {
+      const idx = Math.floor(rng() * arr.length) % arr.length;
+      return arr[idx] || arr[0];
+    }
+    // Deterministic fallback: first element when RNG unavailable
+    return arr[0];
+  };
+  npc = npc || pick(near, ctx.rng);
+
+  // Recruitable follower hire NPCs: bump opens a hire prompt instead of generic chatter/shop.
+  try {
+    if (npc && npc._recruitCandidate && npc._recruitFollowerId) {
+      const FR =
+        ctx.FollowersRuntime ||
+        getMod(ctx, "FollowersRuntime") ||
+        (typeof window !== "undefined" ? window.FollowersRuntime : null);
+      const UIO =
+        ctx.UIOrchestration ||
+        getMod(ctx, "UIOrchestration") ||
+        (typeof window !== "undefined" ? window.UIOrchestration : null);
+
+      if (
+        FR &&
+        typeof FR.canHireFollower === "function" &&
+        typeof FR.hireFollowerFromArchetype === "function"
+      ) {
+        const archetypeId = String(npc._recruitFollowerId || "");
+        if (archetypeId) {
+          const check = FR.canHireFollower(ctx, archetypeId);
+          if (!check.ok) {
+            try {
+              if (ctx.log && check.reason) ctx.log(check.reason, "info");
+            } catch (_) {}
+            return true;
+          }
+
+          // Determine hire cost in gold; simple flat price for now.
+          const hirePrice = 80;
+          // Inspect player gold stack in inventory.
+          let goldObj = null;
+          let goldAmt = 0;
+          try {
+            const inv = ctx.player && Array.isArray(ctx.player.inventory)
+              ? ctx.player.inventory
+              : [];
+            for (let i = 0; i < inv.length; i++) {
+              const it = inv[i];
+              if (it && it.kind === "gold") {
+                goldObj = it;
+                goldAmt =
+                  typeof it.amount === "number" ? (it.amount | 0) : 0;
+                break;
+              }
+            }
+          } catch (_) {}
+          const canAfford = goldAmt >= hirePrice;
+
+          let label = npc.name || "Follower";
+          let archetypeName = "";
+          try {
+            if (typeof FR.getFollowerArchetypes === "function") {
+              const defs = FR.getFollowerArchetypes(ctx) || [];
+              for (let i = 0; i < defs.length; i++) {
+                const d = defs[i];
+                if (!d || !d.id) continue;
+                if (String(d.id) === archetypeId) {
+                  archetypeName = d.name || "";
+                  break;
+                }
+              }
+            }
+          } catch (_) {}
+          if (!label && archetypeName) label = archetypeName;
+
+          const priceStr = `${hirePrice} gold`;
+          const prompt = canAfford
+            ? `${label}: "I can travel with you for ${priceStr}, if you'll have me."`
+            : `${label}: "I'd ask for ${priceStr}, but you don't seem to have that right now."`;
+
+          const onOk = () => {
+            try {
+              if (!canAfford) {
+                if (ctx.log) ctx.log("You don't have enough gold to hire them.", "warn");
+                return;
+              }
+              // Re-check caps before finalizing, in case something changed since prompt.
+              const freshCheck = FR.canHireFollower(ctx, archetypeId);
+              if (!freshCheck || !freshCheck.ok) {
+                if (ctx.log && freshCheck && freshCheck.reason) {
+                  ctx.log(freshCheck.reason, "info");
+                }
+                return;
+              }
+              // Ensure a gold stack exists.
+              if (!goldObj) {
+                const inv = ctx.player && Array.isArray(ctx.player.inventory)
+                  ? ctx.player.inventory
+                  : (ctx.player.inventory = []);
+                goldObj = { kind: "gold", amount: 0, name: "gold" };
+                inv.push(goldObj);
+              }
+              goldObj.amount = (goldObj.amount | 0) - hirePrice;
+
+              const ok = FR.hireFollowerFromArchetype(ctx, archetypeId);
+              if (ok) {
+                // Remove this hire NPC from town after they agree to join.
+                try {
+                  if (Array.isArray(ctx.npcs)) {
+                    for (let i = ctx.npcs.length - 1; i >= 0; i--) {
+                      if (ctx.npcs[i] === npc) {
+                        ctx.npcs.splice(i, 1);
+                        break;
+                      }
+                    }
+                  }
+                  if (
+                    ctx.occupancy &&
+                    typeof ctx.occupancy.clearNPC === "function"
+                  ) {
+                    ctx.occupancy.clearNPC(npc.x | 0, npc.y | 0);
+                  }
+                } catch (_) {}
+                if (ctx.log) {
+                  ctx.log(
+                    `${label} agrees to join you for ${priceStr}.`,
+                    "good"
+                  );
+                }
+              } else if (ctx.log) {
+                ctx.log("They cannot join you right now.", "info");
+              }
+              try {
+                if (typeof ctx.updateUI === "function") ctx.updateUI();
+              } catch (_) {}
+            } catch (_) {}
+          };
+          const onCancel = () => {
+            try {
+              if (ctx.log) {
+                ctx.log(`${label}: "Maybe another time."`, "info");
+              }
+            } catch (_) {}
+          };
+
+          if (UIO && typeof UIO.showConfirm === "function") {
+            UIO.showConfirm(ctx, prompt, null, onOk, onCancel);
+          } else {
+            onOk();
+          }
+          return true;
+        }
+      }
+    }
+  } catch (_) {}
+
+  // Followers in town: bump opens follower inspect panel instead of generic chatter/shop.
+  try {
+    if (npc && npc._isFollower && npc._followerId) {
+      const UIO =
+        ctx.UIOrchestration ||
+        getMod(ctx, "UIOrchestration") ||
+        (typeof window !== "undefined" ? window.UIOrchestration : null);
+      if (UIO && typeof UIO.showFollower === "function") {
+        // Build a minimal runtime follower stub; UIOrchestration will enrich it using
+        // player.followers and followers.json.
+        const followers =
+          ctx.player && Array.isArray(ctx.player.followers)
+            ? ctx.player.followers
+            : [];
+        const rec =
+          followers.find(f => f && f.id === npc._followerId) || null;
+        const runtime = {
+          _isFollower: true,
+          _followerId: npc._followerId,
+          id: npc._followerId,
+          type: npc._followerId,
+          name: npc.name || (rec && rec.name) || "Follower",
+          faction: "guard",
+          level:
+            rec && typeof rec.level === "number" ? rec.level : 1,
+          hp: rec && typeof rec.hp === "number" ? rec.hp : undefined,
+          maxHp:
+            rec && typeof rec.maxHp === "number" ? rec.maxHp : undefined,
+        };
+        UIO.showFollower(ctx, runtime);
+      } else if (ctx.log) {
+        ctx.log(
+          `${npc.name || "Follower"}: ${
+            Array.isArray(npc.lines) && npc.lines[0]
+              ? npc.lines[0]
+              : "I'm with you."
+          }`,
+          "info"
+        );
+      }
+      return true;
+    }
+  } catch (_) {}
+
+  const lines =
+    Array.isArray(npc.lines) && npc.lines.length
+      ? npc.lines
+      : ["Hey!", "Watch it!", "Careful there."];
+  let line = pick(lines, ctx.rng);
+  // Normalize keeper lines for Inn: always open, avoid misleading schedule phrases
+  try {
+    const shopRef = npc && (npc._shopRef || null);
+    if (
+      shopRef &&
+      String(shopRef.type || "").toLowerCase() === "inn" &&
+      !!shopRef.alwaysOpen
+    ) {
+      const s = String(line || "").toLowerCase();
+      if (
+        s.includes("open") ||
+        s.includes("closed") ||
+        s.includes("schedule") ||
+        s.includes("dawn") ||
+        s.includes("dusk")
+      ) {
+        line = "We're open day and night.";
+      }
+    }
+  } catch (_) {}
+  ctx.log &&
+    ctx.log(`${npc.name || "Villager"}: ${line}`, "info");
+
+  // Only shopkeepers can open shops; villagers should not trigger trading.
+  const isKeeper = !!(npc && (npc.isShopkeeper || npc._shopRef));
+
+  // Determine if keeper is at their shop:
+  // - on the door tile
+  // - adjacent to the door (preferred spawn avoids blocking the door itself)
+  // - inside the building
+  function isKeeperAtShop(n, shop) {
+    if (!n || !shop) return false;
+    const atDoor = n.x === shop.x && n.y === shop.y;
+    let inside = false;
+    try {
+      const b = shop.building || null;
+      if (b) {
+        inside =
+          n.x > b.x &&
+          n.x < b.x + b.w - 1 &&
+          n.y > b.y &&
+          n.y < b.y + b.h - 1;
+      }
+    } catch (_) {}
+    // Adjacent to door (outside or just inside) counts as being "at" the shop for interaction.
+    // Accept both cardinal and diagonal adjacency to the door (Chebyshev distance <= 1).
+    const dx = Math.abs(n.x - shop.x),
+      dy = Math.abs(n.y - shop.y);
+    const nearDoor = dx + dy === 1 || Math.max(dx, dy) === 1;
+    return atDoor || inside || nearDoor;
+  }
+
+  // Helper to open a shop reference (if open), showing schedule when closed
+  function tryOpenShopRef(shopRef, sourceNpc) {
+    try {
+      const SS =
+        ctx.ShopService ||
+        (typeof window !== "undefined" ? window.ShopService : null);
+      const openNow =
+        SS && typeof SS.isShopOpenNow === "function"
+          ? SS.isShopOpenNow(ctx, shopRef)
+          : false;
+      const sched =
+        SS && typeof SS.shopScheduleStr === "function"
+          ? SS.shopScheduleStr(shopRef)
+          : "";
+      if (openNow) {
+        try {
+          const UIO =
+            ctx.UIOrchestration ||
+            (typeof window !== "undefined"
+              ? window.UIOrchestration
+              : null);
+          if (UIO && typeof UIO.showShop === "function") {
+            UIO.showShop(ctx, sourceNpc || npc);
+          }
+        } catch (_) {}
+        // UIOrchestration.showShop schedules draw when opening; no manual draw needed
+        return true;
+      } else {
+        ctx.log &&
+          ctx.log(
+            `The ${shopRef.name || "shop"} is closed. ${sched}`,
+            "warn"
+          );
+      }
+    } catch (_) {}
+    return false;
+  }
+
+  if (isKeeper) {
+    try {
+      let shopRef = npc._shopRef || null;
+
+      // Fallback: if keeper lacks a shopRef (e.g., after re-entering town from persistence),
+      // attach the nearest shop by proximity to door or interior.
+      if (!shopRef && Array.isArray(ctx.shops)) {
+        let best = null;
+        let bestScore = Infinity;
+        for (const s of ctx.shops) {
+          if (!s) continue;
+          let inside = false;
+          try {
+            const b = s.building || null;
+            if (b) {
+              inside =
+                npc.x > b.x &&
+                npc.x < b.x + b.w - 1 &&
+                npc.y > b.y &&
+                npc.y < b.y + b.h - 1;
+            }
+          } catch (_) {}
+          const dx = Math.abs(npc.x - s.x),
+            dy = Math.abs(npc.y - s.y);
+          const nearDoor = dx + dy === 1 || Math.max(dx, dy) <= 1;
+          const close = dx + dy <= 2;
+          if (inside || nearDoor || close) {
+            const score = inside ? 0 : nearDoor ? 1 : dx + dy;
+            if (score < bestScore) {
+              bestScore = score;
+              best = s;
+            }
+          }
+        }
+        if (best) {
+          npc._shopRef = best;
+          shopRef = best;
+        }
+      }
+
+      if (shopRef) {
+        // Inn: always open and interactable anywhere inside — open immediately on bump
+        const isInn = String(shopRef.type || "").toLowerCase() === "inn";
+        if (isInn) {
+          tryOpenShopRef(shopRef, npc);
+          return true;
+        }
+
+        if (isKeeperAtShop(npc, shopRef)) {
+          // Keeper at or near the shop door (or inside) — open directly if hours allow
+          tryOpenShopRef(shopRef, npc);
+        } else {
+          // Fallback: if the shop is currently open, allow trading when bumping the keeper anywhere
+          // This makes trading more forgiving when the keeper is on their way or patrolling nearby.
+          const SS =
+            ctx.ShopService ||
+            (typeof window !== "undefined" ? window.ShopService : null);
+          const openNow =
+            SS && typeof SS.isShopOpenNow === "function"
+              ? SS.isShopOpenNow(ctx, shopRef)
+              : false;
+          if (openNow) {
+            tryOpenShopRef(shopRef, npc);
+          } else {
+            const sched =
+              SS && typeof SS.shopScheduleStr === "function"
+                ? SS.shopScheduleStr(shopRef)
+                : "";
+            ctx.log &&
+              ctx.log(
+                `${npc.name || "Shopkeeper"} is away from the ${
+                  shopRef.name || "shop"
+                }. ${sched ? "(" + sched + ")" : ""}`,
+                "info"
+              );
+          }
+        }
+      } else {
+        // No shop resolved; log and return
+        ctx.log &&
+          ctx.log(
+            `${npc.name || "Shopkeeper"}: shop not found nearby.`,
+            "warn"
+          );
+      }
+    } catch (_) {}
+    return true;
+  }
+
+  // Do not auto-open shops when bumping non-keepers, even if near a door.
+  return true;
+}
